Refactor: cctv service
