+++ detection/src/test/java/com/abnormal/detection/service/CctvServiceTest.java
@@ -1,6 +1,7 @@
 package com.abnormal.detection.service;
 
 
+import com.abnormal.detection.service.cctv.CctvService;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -11,7 +12,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.Optional;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
